+++ src/collective/plausible/tests/test_robot.py
@@ -1,5 +1,7 @@
 # -*- coding: utf-8 -*-
-from collective.plausible.testing import COLLECTIVE_PLAUSIBLE_ACCEPTANCE_TESTING  # noqa: E501
+from collective.plausible.testing import (
+    COLLECTIVE_PLAUSIBLE_ACCEPTANCE_TESTING,
+)  # noqa: E501
 from plone.app.testing import ROBOT_TEST_LEVEL
 from plone.testing import layered