+++ src/collective/plausible/tests/test_setup.py
@@ -1,6 +1,8 @@
 # -*- coding: utf-8 -*-
 """Setup tests for this package."""
-from collective.plausible.testing import COLLECTIVE_PLAUSIBLE_INTEGRATION_TESTING  # noqa: E501
+from collective.plausible.testing import (
+    COLLECTIVE_PLAUSIBLE_INTEGRATION_TESTING,
+)  # noqa: E501
 from plone import api
 from plone.app.testing import setRoles
 from plone.app.testing import TEST_USER_ID